requirements.txt
requests
faker
locustio
